incorporating manager into status win
